Add Makefile, and complete OS X platform.
This adds the minimum number of syscalls needed to compile, and the
appropriate C type sizes. Because we’re not linking against libSystem
(and using the LC_UNIXTHREAD command) we use the clang flag `-e _start`
to set the entry point at rust’s unavoidably name-mangled start
function.

This also adds a Makefile capable of building on multiple platforms,
with the caveat that the `-L $(TARGETDIR)/deps` flag must be added to
RUSTCFLAGS for non-native platforms.